+++ who_l3_smart_tools/cli/indicator_testing.py
@@ -3,9 +3,9 @@ import os
 import argparse
 from pathlib import Path
 import pandas as pd
-from who_l3_smart_tools.core.indicator_testing.bundle_generator import BundleGenerator
-from who_l3_smart_tools.core.indicator_testing.data_generator import DataGenerator
-from who_l3_smart_tools.core.indicator_testing.scaffolding_generator import (
+from ..core.indicator_testing.bundle_generator import BundleGenerator
+from ..core.indicator_testing.data_generator import DataGenerator
+from ..core.indicator_testing.scaffolding_generator import (
     ScaffoldingGenerator,
 )
