+++ who_l3_smart_tools/utils/ocl.py
@@ -1,5 +1,7 @@
 import slugify
 
+__all__ = ["dash_preserving_slugify"]
+
 
 def dash_preserving_slugify(text):
     """
